fix(modeling) moved measureEpsilon into measurements
