+++ packages/modeling/src/operations/booleans/retessellate.js
@@ -5,11 +5,21 @@ const poly3 = require('../../geometries/poly3')
 
 const reTesselateCoplanarPolygons = require('./reTesselateCoplanarPolygons')
 
+// Normals are directional vectors of length 0 to 1.0
+// This EPS is derived from a serieas of tests to determine the optimal precision for comparing coplanar polygons,
+// as provided by the sphere primitive at high segmentation
+// This EPS is for 64 bit Number values
+const NEPS = 1e-13
+
+// Compare two normals (unit vectors) for equality.
+const aboutEqualNormals = (a, b) => {
+  return (Math.abs(a[0] - b[0]) <= NEPS && Math.abs(a[1] - b[1]) <= NEPS && Math.abs(a[2] - b[2]) <= NEPS)
+}
+
 const coplanar = (plane1, plane2) => {
   // expect the same distance from the origin, within tolerance
   if (Math.abs(plane1[3] - plane2[3]) < 0.00000015) {
-    // expect a zero (0) angle between the normals
-    if (vec3.angle(plane1, plane2) === 0) return true
+    return aboutEqualNormals(plane1, plane2)
   }
   return false
 }
