+++ packages/modeling/src/operations/booleans/subtractGeom2.js
@@ -2,11 +2,11 @@ const flatten = require('../../utils/flatten')
 
 const geom3 = require('../../geometries/geom3')
 
+const measureEpsilon = require('../../measurements/measureEpsilon')
+
 const fromFakePolygons = require('./fromFakePolygons')
 const to3DWalls = require('./to3DWalls')
 const subtractGeom3 = require('./subtractGeom3')
-
-const measureEpsilon = require('./measureEpsilon')
 
 /*
  * Return a new 2D geometry representing space in the first geometry but
